fix: Update golden tests for integer literal emission
Update expected C# output to use integer literals (0, -1) instead of
double literals (0.0, -1.0) for comparisons with int-returning methods
like Array.length and String.length.

This reflects the correct behavior after the branded-int feature where
integer types are preserved from declarations.

+++ packages/emitter/testcases/real-world/business-logic/business-logic.cs
@@ -72,7 +72,7 @@ namespace TestCases.realworld.businesslogic
         public bool removeItem(string productId)
             {
             var index = global::Tsonic.JSRuntime.Array.findIndex(this.items, (CartItem item) => item.product.id == productId);
-            if (index != -1.0)
+            if (index != (int)(-1.0))
                 {
                 global::Tsonic.JSRuntime.Array.splice(this.items, index, 1.0);
                 return true;
@@ -162,7 +162,7 @@ namespace TestCases.realworld.businesslogic
         public Order createOrder(ShoppingCart cart)
             {
             var items = cart.getItems();
-            if (global::Tsonic.JSRuntime.Array.length(items) == 0.0)
+            if (global::Tsonic.JSRuntime.Array.length(items) == 0)
                 {
                 throw new Error("Cannot create order from empty cart");
                 }
